+++ src/rust/aggregator/sqlx_queries/order_history/get_latest_event_timestamp.sql
@@ -1,0 +1,1 @@
+SELECT * FROM aggregator.order_history_latest_event_timestamp;