+++ src/rust/aggregator/sqlx_queries/order_history/update_last_indexed_timestamp.sql
@@ -1,0 +1,8 @@
+WITH parameters AS (
+    SELECT
+        $1::timestamptz AS "time"
+)
+UPDATE aggregator.order_history_last_indexed_timestamp
+SET
+    "time" = parameters."time"
+FROM parameters