+++ src/terraform/dss-ci-cd/dss/modules/grafana/variables.tf
@@ -1,0 +1,15 @@
+variable "db_conn_str_private_grafana" {}
+
+variable "db_private_ip_and_port" {}
+
+variable "migrations_complete" {}
+
+variable "no_auth_policy_data" {}
+
+variable "region" {}
+
+variable "sql_vpc_connector_id" {}
+
+variable "grafana_admin_password" {}
+
+variable "grafana_public_password" {}