+++ src/terraform/dss-ci-cd/dss/modules/no_auth_policy/main.tf
@@ -1,0 +1,8 @@
+data "google_iam_policy" "no_auth" {
+  binding {
+    role = "roles/run.invoker"
+    members = [
+      "allUsers",
+    ]
+  }
+}