+++ src/terraform/dss-ci-cd/dss/modules/postgrest/variables.tf
@@ -1,0 +1,11 @@
+variable "db_conn_str_private" {}
+
+variable "migrations_complete" {}
+
+variable "no_auth_policy_data" {}
+
+variable "postgrest_max_rows" {}
+
+variable "region" {}
+
+variable "sql_vpc_connector_id" {}